Moving more data files into core-models module.
