+++ app/src/main/java/com/adammcneilly/pocketleague/core/data/remote/smashgg/SmashGGModelMapper.kt
@@ -1,10 +1,16 @@
 package com.adammcneilly.pocketleague.core.data.remote.smashgg
 
-import com.adammcneilly.pocketleague.bracket.domain.models.BracketType
-import com.adammcneilly.pocketleague.core.domain.models.Player
-import com.adammcneilly.pocketleague.core.domain.models.Team
+import com.adammcneilly.pocketleague.core.models.BracketType
+import com.adammcneilly.pocketleague.core.models.EventOverview
+import com.adammcneilly.pocketleague.core.models.EventSet
+import com.adammcneilly.pocketleague.core.models.PhaseDetail
+import com.adammcneilly.pocketleague.core.models.PhaseOverview
+import com.adammcneilly.pocketleague.core.models.Player
+import com.adammcneilly.pocketleague.core.models.SetSlot
+import com.adammcneilly.pocketleague.core.models.Standings
+import com.adammcneilly.pocketleague.core.models.StandingsPlacement
+import com.adammcneilly.pocketleague.core.models.Team
 import com.adammcneilly.pocketleague.event.data.remote.ApolloBracketType
-import com.adammcneilly.pocketleague.eventoverview.domain.models.EventOverview
 import com.adammcneilly.pocketleague.fragment.EventEntrantFragment
 import com.adammcneilly.pocketleague.fragment.EventOverviewFragment
 import com.adammcneilly.pocketleague.fragment.EventPlayerFragment
@@ -13,15 +19,7 @@ import com.adammcneilly.pocketleague.fragment.PhaseDetailFragment
 import com.adammcneilly.pocketleague.fragment.PhaseGroupFragment
 import com.adammcneilly.pocketleague.fragment.SetSlotFragment
 import com.adammcneilly.pocketleague.fragment.StandingsPlacementFragment
-import com.adammcneilly.pocketleague.phase.domain.models.PhaseDetail
-import com.adammcneilly.pocketleague.phase.domain.models.PhaseOverview
-import com.adammcneilly.pocketleague.set.domain.models.EventSet
-import com.adammcneilly.pocketleague.set.domain.models.SetSlot
-import com.adammcneilly.pocketleague.standings.domain.models.Standings
-import com.adammcneilly.pocketleague.standings.domain.models.StandingsPlacement
 import com.apollographql.apollo.api.BigDecimal
-import java.time.Instant
-import java.time.ZoneOffset
 import javax.inject.Inject
 
 /**
@@ -33,7 +31,6 @@ class SmashGGModelMapper @Inject constructor() {
      */
     fun eventOverviewFragmentToEventOverview(eventOverview: EventOverviewFragment?): EventOverview {
         val startSeconds = (eventOverview?.fragments?.eventSummaryFragment?.startAt as BigDecimal).toLong()
-        val startDate = Instant.ofEpochSecond(startSeconds).atOffset(ZoneOffset.UTC)
 
         return EventOverview(
             name = eventOverview.fragments.eventSummaryFragment.name.orEmpty(),
@@ -48,7 +45,7 @@ class SmashGGModelMapper @Inject constructor() {
                     phase.phaseOrder
                 }
                 .orEmpty(),
-            startDate = startDate.toZonedDateTime(),
+            startDateEpochSeconds = startSeconds,
             standings = Standings(
                 placements = eventOverview.standings
                     ?.nodes
