+++ app/src/main/java/com/adammcneilly/pocketleague/core/utils/DateTimeHelper.kt
@@ -1,7 +1,6 @@
 package com.adammcneilly.pocketleague.core.utils
 
 import java.time.LocalDateTime
-import java.time.ZonedDateTime
 
 /**
  * A collection of helper functions for dealing with date and time objects.
@@ -13,7 +12,7 @@ interface DateTimeHelper {
     fun getSeriesDayTimeString(date: LocalDateTime): String
 
     /**
-     * Given a [date], convert it to a user friendly string to show the day of this event.
+     * Given some [epochSeconds], convert it to a user friendly string to show the day of this event.
      */
-    fun getEventDayString(date: ZonedDateTime): String
+    fun getEventDayString(epochSeconds: Long): String
 }
