+++ app/src/main/java/com/adammcneilly/pocketleague/core/utils/DateUtils.kt
@@ -1,7 +1,8 @@
 package com.adammcneilly.pocketleague.core.utils
 
+import java.time.Instant
 import java.time.LocalDateTime
-import java.time.ZonedDateTime
+import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
 
 /**
@@ -16,7 +17,12 @@ object DateUtils : DateTimeHelper {
         return DateTimeFormatter.ofPattern(SERIES_DAY_TIME_FORMAT).format(date)
     }
 
-    override fun getEventDayString(date: ZonedDateTime): String {
-        return DateTimeFormatter.ofPattern(EVENT_DAY_FORMAT).format(date)
+    override fun getEventDayString(epochSeconds: Long): String {
+        val zonedDateTime = Instant
+            .ofEpochSecond(epochSeconds)
+            .atOffset(ZoneOffset.UTC)
+            .toZonedDateTime()
+
+        return DateTimeFormatter.ofPattern(EVENT_DAY_FORMAT).format(zonedDateTime)
     }
 }
