+++ app/src/main/java/com/adammcneilly/pocketleague/eventoverview/domain/state/EventOverviewStateMutator.kt
@@ -2,20 +2,20 @@
 
 package com.adammcneilly.pocketleague.eventoverview.domain.state
 
-import com.adammcneilly.pocketleague.bracket.domain.models.BracketType
 import com.adammcneilly.pocketleague.core.data.Result
-import com.adammcneilly.pocketleague.core.domain.models.Player
+import com.adammcneilly.pocketleague.core.models.BracketType
+import com.adammcneilly.pocketleague.core.models.EventOverview
+import com.adammcneilly.pocketleague.core.models.PhaseOverview
+import com.adammcneilly.pocketleague.core.models.Player
+import com.adammcneilly.pocketleague.core.models.Standings
+import com.adammcneilly.pocketleague.core.models.StandingsPlacement
 import com.adammcneilly.pocketleague.core.ui.UIImage
 import com.adammcneilly.pocketleague.core.ui.UIText
 import com.adammcneilly.pocketleague.core.utils.DateTimeHelper
-import com.adammcneilly.pocketleague.eventoverview.domain.models.EventOverview
 import com.adammcneilly.pocketleague.eventoverview.domain.usecases.FetchEventOverviewUseCase
 import com.adammcneilly.pocketleague.eventoverview.ui.EventOverviewDisplayModel
 import com.adammcneilly.pocketleague.eventoverview.ui.EventOverviewViewState
-import com.adammcneilly.pocketleague.phase.domain.models.PhaseOverview
 import com.adammcneilly.pocketleague.phase.ui.PhaseDisplayModel
-import com.adammcneilly.pocketleague.standings.domain.models.Standings
-import com.adammcneilly.pocketleague.standings.domain.models.StandingsPlacement
 import com.adammcneilly.pocketleague.standings.ui.StandingsDisplayModel
 import com.adammcneilly.pocketleague.standings.ui.StandingsPlacementDisplayModel
 import com.tunjid.mutator.Mutation
@@ -147,7 +147,7 @@ private fun EventOverview.toDisplayModel(
                 },
             )
         },
-        startDate = dateTimeHelper.getEventDayString(this.startDate),
+        startDate = dateTimeHelper.getEventDayString(this.startDateEpochSeconds),
         standings = this.standings.toDisplayModel(),
     )
 }
@@ -171,7 +171,7 @@ private fun PhaseOverview.toDisplayModel(
 
 private fun Standings.toDisplayModel(): StandingsDisplayModel {
     return StandingsDisplayModel(
-        placements = this.placements.map(StandingsPlacement::toDisplayModel)
+        placements = this.placements.map(com.adammcneilly.pocketleague.core.models.StandingsPlacement::toDisplayModel)
     )
 }
