+++ app/src/main/java/com/adammcneilly/pocketleague/eventoverview/domain/usecases/FetchEventOverviewUseCaseImpl.kt
@@ -1,8 +1,8 @@
 package com.adammcneilly.pocketleague.eventoverview.domain.usecases
 
 import com.adammcneilly.pocketleague.core.data.Result
+import com.adammcneilly.pocketleague.core.models.EventOverview
 import com.adammcneilly.pocketleague.event.data.EventService
-import com.adammcneilly.pocketleague.eventoverview.domain.models.EventOverview
 import javax.inject.Inject
 
 /**
