+++ app/src/main/java/com/adammcneilly/pocketleague/phase/data/PhaseService.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.pocketleague.phase.data
 
 import com.adammcneilly.pocketleague.core.data.Result
-import com.adammcneilly.pocketleague.phase.domain.models.PhaseDetail
+import com.adammcneilly.pocketleague.core.models.PhaseDetail
 
 /**
  * The data contract to fetch information about a phase.
