+++ app/src/test/java/com/adammcneilly/pocketleague/eventsummary/domain/state/EventSummaryListStateMutatorTest.kt
@@ -13,8 +13,6 @@ import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.TestCoroutineScope
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
-import java.time.Instant
-import java.time.ZoneOffset
 
 class EventSummaryListStateMutatorTest {
     private val fetchUpcomingEventsUseCase = FakeFetchUpcomingEventsUseCase()
@@ -44,11 +42,7 @@ class EventSummaryListStateMutatorTest {
 
         // Mocks
         fetchUpcomingEventsUseCase.mockResult = fakeEventListResult
-        val startDate = Instant
-            .ofEpochSecond(fakeEvent.startDateEpochSeconds)
-            .atOffset(ZoneOffset.UTC)
-            .toZonedDateTime()
-        dateTimeHelper.mockEventDayStringForDate(startDate, fakeEventDateString)
+        dateTimeHelper.mockEventDayStringForDate(fakeEvent.startDateEpochSeconds, fakeEventDateString)
 
         // Expectations
         val initialState = EventSummaryListViewState()
