+++ app/src/test/java/com/adammcneilly/pocketleague/player/ui/PlayerDisplayModelTest.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.pocketleague.player.ui
 
-import com.adammcneilly.pocketleague.core.domain.models.Player
+import com.adammcneilly.pocketleague.core.models.Player
 import com.adammcneilly.pocketleague.core.ui.FakeFlagResProvider
 import com.adammcneilly.pocketleague.core.ui.UIImage
 import com.google.common.truth.Truth.assertThat
