+++ app/src/test/java/com/adammcneilly/pocketleague/teamdetail/ui/TeamDetailDisplayModelTest.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.pocketleague.teamdetail.ui
 
-import com.adammcneilly.pocketleague.core.domain.models.Player
-import com.adammcneilly.pocketleague.core.domain.models.Team
+import com.adammcneilly.pocketleague.core.models.Player
+import com.adammcneilly.pocketleague.core.models.Team
 import com.adammcneilly.pocketleague.core.ui.FakeFlagResProvider
 import com.adammcneilly.pocketleague.core.ui.UIImage
 import com.adammcneilly.pocketleague.player.ui.toDisplayModel
