--- core-models/src/androidMain/kotlin/com/adammcneilly/pocketleague/core/models/Platform.kt
@@ -1,5 +1,0 @@
-package com.adammcneilly.pocketleague.core.models
-
-actual class Platform actual constructor() {
-    actual val platform: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
