--- core-models/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/models/Greeting.kt
@@ -1,7 +1,0 @@
-package com.adammcneilly.pocketleague.core.models
-
-class Greeting {
-    fun greeting(): String {
-        return "Hello, ${Platform().platform}!"
-    }
-}
