--- core-models/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/models/Platform.kt
@@ -1,5 +1,0 @@
-package com.adammcneilly.pocketleague.core.models
-
-expect class Platform() {
-    val platform: String
-}
