+++ core-models/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/models/StandingsPlacement.kt
@@ -1,6 +1,4 @@
-package com.adammcneilly.pocketleague.standings.domain.models
-
-import com.adammcneilly.pocketleague.core.domain.models.Team
+package com.adammcneilly.pocketleague.core.models
 
 /**
  * Represents the [placement] within standings for a given [team].
