--- core-models/src/iosMain/kotlin/com/adammcneilly/pocketleague/core/models/Platform.kt
@@ -1,7 +1,0 @@
-package com.adammcneilly.pocketleague.core.models
-
-import platform.UIKit.UIDevice
-
-actual class Platform actual constructor() {
-    actual val platform: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
